requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyserial>=3.5
pydantic>=2.0.0
pydantic-settings>=2.0.0
